requirements.txt
python-telegram-bot>=20.0
APScheduler>=3.10
python-dotenv>=1.0
python-telegram-bot

